[Distiller] Move temperature reading to the use case.

+++ distiller/app/src/main/java/com/github/wmarkow/distiller/ui/presenter/DistillerFakeDataPresenter.java
@@ -18,11 +18,11 @@ public class DistillerFakeDataPresenter extends DistillerDataPresenter {
     public void readDistillerData() {
         DistillerData dd = new DistillerData();
         dd.deviceUpTimeMillis = System.currentTimeMillis();
-        dd.coldWaterTemp = (float) (Math.random() * 1) + 15f;
-        dd.hotWaterTemp = (float) (Math.random() * 1) + 76f;
-        dd.boilerTemp = (float) (Math.random() * 0.2) + 91.5f;
-        dd.headerTemp = (float) (Math.random() * 0.2) + 80.8f;
-        dd.waterRpm = (float)(Math.random() * 50) + 1800f;
+        dd.coldWaterTemp = (Math.random() * 1) + 15f;
+        dd.hotWaterTemp = (Math.random() * 1) + 76f;
+        dd.boilerTemp = (Math.random() * 0.2) + 91.5f;
+        dd.headerTemp = (Math.random() * 0.2) + 80.8f;
+        dd.waterRpm = (Math.random() * 50) + 1800f;
 
         //getDistillerDataView().showNewDistillerData(dd);
     }
